requirements.txt
tensorflow-cpu==2.6.0
gunicorn==20.1.0
Flask==2.0.2
h5py==3.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.19.5
opencv-python==4.5.4.58
Werkzeug==2.0.2